requirements-dev.txt
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2
Pillow==6.0.0
python-decouple==3.1
pytz==2018.9
sqlparse==0.3.0
static3==0.7.0
django-bootstrap-form==3.4
django-crispy-forms==1.7.2
